refactor: extract lane-id to a new component
