+++ scopes/lanes/lane-id/index.ts
@@ -1,0 +1,3 @@
+export { LaneId, DEFAULT_LANE, LANE_REMOTE_DELIMITER } from './lane-id';
+export { RemoteLaneId } from './remote-lane-id';
+export { LocalLaneId } from './local-lane-id';
