+++ scopes/lanes/lane-id/local-lane-id.ts
@@ -1,0 +1,10 @@
+import { LaneId } from './lane-id';
+
+export class LocalLaneId extends LaneId {
+  constructor({ name }: { name: string }) {
+    super({ name });
+  }
+  static from(name: string): LocalLaneId {
+    return new LocalLaneId({ name });
+  }
+}
