+++ scopes/lanes/lanes/merge-lanes.ts
@@ -1,9 +1,8 @@
 import { BitError } from '@teambit/bit-error';
 import { BitId } from '@teambit/legacy-bit-id';
-import { DEFAULT_LANE } from '@teambit/legacy/dist/constants';
 import { Consumer } from '@teambit/legacy/dist/consumer';
 import { ApplyVersionResults, MergeStrategy } from '@teambit/legacy/dist/consumer/versions-ops/merge-version';
-import LaneId, { RemoteLaneId } from '@teambit/legacy/dist/lane-id/lane-id';
+import { LaneId, RemoteLaneId, DEFAULT_LANE } from '@teambit/lane-id';
 import { Lane } from '@teambit/legacy/dist/scope/models';
 import { Tmp } from '@teambit/legacy/dist/scope/repositories';
 import { MergingMain } from '@teambit/merging';
