+++ src/api/scope/lib/cat-lane.ts
@@ -1,5 +1,5 @@
+import { LaneId } from '@teambit/lane-id';
 import GeneralError from '../../../error/general-error';
-import LaneId from '../../../lane-id/lane-id';
 import { loadScope, Scope } from '../../../scope';
 
 export default async function catLane(name: string) {
