+++ src/scope/scope.ts
@@ -2,6 +2,7 @@ import fs from 'fs-extra';
 import mapSeries from 'p-map-series';
 import * as pathLib from 'path';
 import R from 'ramda';
+import { LaneId, RemoteLaneId } from '@teambit/lane-id';
 import semver from 'semver';
 import { Analytics } from '../analytics/analytics';
 import { BitId, BitIds } from '../bit-id';
@@ -30,7 +31,6 @@ import Consumer from '../consumer/consumer';
 import SpecsResults from '../consumer/specs-results';
 import { SpecsResultsWithComponentId } from '../consumer/specs-results/specs-results';
 import GeneralError from '../error/general-error';
-import LaneId, { RemoteLaneId } from '../lane-id/lane-id';
 import logger from '../logger/logger';
 import getMigrationVersions, { MigrationResult } from '../migration/migration-helper';
 import { currentDirName, first, pathHasAll, propogateUntil, readDirSyncIgnoreDsStore } from '../utils';
@@ -108,8 +108,8 @@ export type OnTagOpts = {
   isSnap?: boolean;
 };
 export type IsolateComponentsOptions = {
-  packageManagerConfigRootDir?: string
-}
+  packageManagerConfigRootDir?: string;
+};
 export type OnTagFunc = (
   components: Component[],
   options: OnTagOpts,
